Update to use the new LaunchLibrary 2.0+ API
The older API is being depricated and will no longer be used.

This commit changes the API used and fixes any minor issues introduced
from the differences between the API versions.

Note: the video stream URL will need to be verified, as no current
upcoming launches have this data, so it cannot be validated at this
time.